+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
  * added `RawArgs::getScriptName()` and `Args::getScriptName()`
  * enabled nice exception rendering for exceptions thrown before the IO is created
+ * made it possible to pass a callable to `ConsoleApplication::__construct()`
 
 * 1.0.0-beta (2015-03-19)